Mostly cleaned, all tested
Tested all nlmtst-functions, and made adjustments where necessary. Main changes:
get() --> get_nl(), to avoid confusing Matlab. The internal get()-function has changed over time and no longer works the way the toolbox expects.
Very few set() calls had to be changed to set_nl(), but this turned out to be crucial for some. Interestingly, while seemling all get() needed to be changed, some set() actually requires the internal version.
Auto-indented a lot of files and removed pointless (pun intended) semi-colons and commas in several functions.
Added etvc_mex to the src-folder where other .c files are mexed, and added it to the win64 mexer. This allows the TVM object to work properly. Also cleaned up the .c file of etvc_mex, because it was only configured for non-ANSI compilers (and gave conflicts with Matlab's internal definitions).
The only function that remains broken is WSERIES(), because it calls a laguerre() function that does not exist in Matlab or in the toolbox.
Changed the way the nlmtst(fresp) function works: Matlab 2018b tries to evaluate fresp as a variable first, instead of as a function, so an feval was needed.
Many other minor best-practice upgrades as recommended by Matlab (e.g. isstr() --> ischar())